rtc_pl031: Don't init counter to 0, a little more logging
qemu initialises the RTC counter to the unix time on the host.
Setting the counter to 0 makes time go backwards and this is potentially
AFTER we've already read the RTC counter to print the time for log
messages. This causes an ASSERT failure. The fix is to use whatever qemu
initialised as our initial system time.